fix: 유저 삭제 인증 추가
- 유저 삭제시 로그인한 유저와 동일한 유저만 삭제 가능하게 변경

+++ src/main/java/com/scheduleplus/user/service/UserService.java
@@ -50,9 +50,18 @@ public class UserService {
     }
 
     @Transactional
-    public void delete(Long userId) {
-        getUser(userId);
-        userRepository.deleteById(userId);
+    public void delete(Long userId, HttpSession session) {
+        SessionValue sessionValue = (SessionValue) session.getAttribute("sessionId");
+        if (sessionValue == null) {
+            throw new IllegalArgumentException("로그인이 필요한 작업입니다.");
+        }
+        boolean isMatch = userId.equals(sessionValue.getUserId());
+        if (isMatch) {
+            userRepository.deleteById(userId);
+        } else {
+            throw new IllegalArgumentException("본인 유저정보만 삭제가 가능합니다.");
+        }
+
     }
 
     @Transactional
